Revert "refactor:Make separate packages for all integration #37 (#39)"
This reverts commit e55ae74e403e4acd158d98ee20e7417210df7c4d.
