+++ integrations/doc_test.go
@@ -12,14 +12,7 @@ import (
 
 	"github.com/bnkamalesh/webgo/v4"
 	"github.com/gin-gonic/gin"
-
-	"github.com/keploy/go-sdk/integrations/echo"
-	"github.com/keploy/go-sdk/integrations/gin"
-	"github.com/keploy/go-sdk/integrations/grpc"
-	"github.com/keploy/go-sdk/integrations/http-client"
-	"github.com/keploy/go-sdk/integrations/mongo"
-	"github.com/keploy/go-sdk/integrations/webgo"
-
+	"github.com/keploy/go-sdk/integrations"
 	"github.com/keploy/go-sdk/keploy"
 	"github.com/labstack/echo/v4"
 	"go.mongodb.org/mongo-driver/bson"
@@ -41,7 +34,7 @@ func ExampleNewMongoCollection() {
 		log.Fatal(err)
 	}
 	fmt.Println("Connected to MongoDB!")
-	var collection *Kmongo.MongoCollection
+	var collection *integrations.MongoCollection
 	result, err := collection.InsertOne(context.TODO(), bson.D{{"x", 1}})
 	if err != nil {
 		log.Fatal(err)
@@ -51,8 +44,8 @@ func ExampleNewMongoCollection() {
 
 func ExampleMongoSingleResult_Err() {
 	var (
-		sr         *Kmongo.MongoSingleResult
-		collection *Kmongo.MongoCollection
+		sr         *integrations.MongoSingleResult
+		collection *integrations.MongoCollection
 	)
 	filter := bson.M{"name": "Ash"}
 	findOneOpts := options.FindOne()
@@ -67,8 +60,8 @@ func ExampleMongoSingleResult_Err() {
 
 func ExampleMongoSingleResult_Decode() {
 	var (
-		sr         *Kmongo.MongoSingleResult
-		collection *Kmongo.MongoCollection
+		sr         *integrations.MongoSingleResult
+		collection *integrations.MongoCollection
 	)
 	filter := bson.M{"name": "Ash"}
 	var result bson.D
@@ -85,8 +78,8 @@ func ExampleMongoSingleResult_Decode() {
 
 func ExampleMongoCursor_Err() {
 	var (
-		collection *Kmongo.MongoCollection
-		cur        *Kmongo.MongoCursor
+		collection *integrations.MongoCollection
+		cur        *integrations.MongoCursor
 		err        error
 	)
 	filter := bson.M{"name": "Misty"}
@@ -101,8 +94,8 @@ func ExampleMongoCursor_Err() {
 
 func ExampleMongoCursor_Next() {
 	var (
-		collection *Kmongo.MongoCollection
-		cur        *Kmongo.MongoCursor
+		collection *integrations.MongoCollection
+		cur        *integrations.MongoCursor
 		err        error
 	)
 	filter := bson.M{"name": "Misty"}
@@ -119,8 +112,8 @@ func ExampleMongoCursor_Next() {
 
 func ExampleMongoCursor_TryNext() {
 	var (
-		collection *Kmongo.MongoCollection
-		cur        *Kmongo.MongoCursor
+		collection *integrations.MongoCollection
+		cur        *integrations.MongoCursor
 		err        error
 	)
 	filter := bson.M{"name": "Misty"}
@@ -137,8 +130,8 @@ func ExampleMongoCursor_TryNext() {
 
 func ExampleMongoCursor_Close() {
 	var (
-		collection *Kmongo.MongoCollection
-		cur        *Kmongo.MongoCursor
+		collection *integrations.MongoCollection
+		cur        *integrations.MongoCursor
 		err        error
 	)
 	filter := bson.M{"name": "Misty"}
@@ -153,8 +146,8 @@ func ExampleMongoCursor_Close() {
 
 func ExampleMongoCursor_All() {
 	var (
-		collection *Kmongo.MongoCollection
-		cur        *Kmongo.MongoCursor
+		collection *integrations.MongoCollection
+		cur        *integrations.MongoCursor
 		err        error
 	)
 	filter := bson.M{"name": "Misty"}
@@ -170,8 +163,8 @@ func ExampleMongoCursor_All() {
 
 func ExampleMongoCursor_Decode() {
 	var (
-		collection *Kmongo.MongoCollection
-		cur        *Kmongo.MongoCursor
+		collection *integrations.MongoCollection
+		cur        *integrations.MongoCursor
 		err        error
 	)
 	filter := bson.M{"name": "Misty"}
@@ -194,7 +187,7 @@ func ExampleMongoCursor_Decode() {
 
 func ExampleMongoCollection_InsertOne() {
 	var (
-		collection      *Kmongo.MongoCollection
+		collection      *integrations.MongoCollection
 		err             error
 		insertOneResult *mongo.InsertOneResult
 	)
@@ -210,8 +203,8 @@ func ExampleMongoCollection_InsertOne() {
 
 func ExampleMongoCollection_FindOne() {
 	var (
-		sr         *Kmongo.MongoSingleResult
-		collection *Kmongo.MongoCollection
+		sr         *integrations.MongoSingleResult
+		collection *integrations.MongoCollection
 		err        error
 	)
 	filter := bson.M{"name": "Ash"}
@@ -230,7 +223,7 @@ func ExampleMongoCollection_FindOne() {
 func ExampleMongoCollection_InsertMany() {
 	var (
 		insertManyResult *mongo.InsertManyResult
-		collection       *Kmongo.MongoCollection
+		collection       *integrations.MongoCollection
 		err              error
 	)
 	docs := []interface{}{
@@ -248,8 +241,8 @@ func ExampleMongoCollection_InsertMany() {
 
 func ExampleMongoCollection_Find() {
 	var (
-		collection *Kmongo.MongoCollection
-		cur        *Kmongo.MongoCursor
+		collection *integrations.MongoCollection
+		cur        *integrations.MongoCursor
 		err        error
 	)
 	filter := bson.M{"name": "Misty"}
@@ -273,7 +266,7 @@ func ExampleMongoCollection_Find() {
 func ExampleMongoCollection_UpdateOne() {
 	var (
 		result     *mongo.UpdateResult
-		collection *Kmongo.MongoCollection
+		collection *integrations.MongoCollection
 		err        error
 	)
 	filter := bson.M{"name": "Brock"}
@@ -296,7 +289,7 @@ func ExampleMongoCollection_UpdateOne() {
 func ExampleMongoCollection_UpdateMany() {
 	var (
 		result     *mongo.UpdateResult
-		collection *Kmongo.MongoCollection
+		collection *integrations.MongoCollection
 		err        error
 	)
 	filter := bson.M{"name": "Brock"}
@@ -316,7 +309,7 @@ func ExampleMongoCollection_UpdateMany() {
 func ExampleMongoCollection_DeleteOne() {
 	var (
 		result     *mongo.DeleteResult
-		collection *Kmongo.MongoCollection
+		collection *integrations.MongoCollection
 		err        error
 	)
 	filter := bson.M{"name": "Brock"}
@@ -332,7 +325,7 @@ func ExampleMongoCollection_DeleteOne() {
 func ExampleMongoCollection_DeleteMany() {
 	var (
 		result     *mongo.DeleteResult
-		collection *Kmongo.MongoCollection
+		collection *integrations.MongoCollection
 		err        error
 	)
 	filter := bson.M{"name": "Brock"}
@@ -362,7 +355,7 @@ func ExampleWebGoV4() {
 		ReadTimeout:  15 * time.Second,
 		WriteTimeout: 60 * time.Second,
 	}, []*webgo.Route{})
-	Kwebgo.WebGoV4(k, router)
+	integrations.WebGoV4(k, router)
 	router.Start()
 }
 
@@ -379,7 +372,7 @@ func ExampleEchoV4() {
 		},
 	})
 	// Remember to call integrations.EchoV4 before route handling
-	Kecho.EchoV4(k, e)
+	integrations.EchoV4(k, e)
 	e.GET("/echo", func(c echo.Context) error {
 		return nil
 	})
@@ -399,7 +392,7 @@ func ExampleGinV1() {
 		},
 	})
 	//Call integration.GinV1 before routes handling
-	Kgin.GinV1(k, r)
+	integrations.GinV1(k, r)
 	r.GET("/gin/:color/*type", func(c *gin.Context) {
 		c.JSON(200, gin.H{
 			"message": "pong",
@@ -418,7 +411,7 @@ func ExampleWithClientUnaryInterceptor() {
 			URL: "http://localhost:8081/api",
 		},
 	})
-	conn, err := grpc.Dial("localhost:50051", grpc.WithInsecure(), Kgrpc.WithClientUnaryInterceptor(k))
+	conn, err := grpc.Dial("localhost:50051", grpc.WithInsecure(), integrations.WithClientUnaryInterceptor(k))
 	if err != nil {
 		log.Fatalf("Did not connect : %v", err)
 	}
@@ -436,7 +429,7 @@ func ExampleWithClientStreamInterceptor() {
 		},
 	})
 
-	conn, err := grpc.Dial("localhost:50051", grpc.WithInsecure(), Kgrpc.WithClientStreamInterceptor(k))
+	conn, err := grpc.Dial("localhost:50051", grpc.WithInsecure(), integrations.WithClientStreamInterceptor(k))
 	if err != nil {
 		log.Fatalf("Did not connect : %v", err)
 	}
@@ -450,7 +443,7 @@ func ExampleNewHttpClient() {
 		IdleConnTimeout:    30 * time.Second,
 		DisableCompression: true,
 	}
-	client := Khttp.NewHttpClient(&http.Client{
+	client := integrations.NewHttpClient(&http.Client{
 		Transport: tr,
 	})
 
@@ -467,7 +460,7 @@ func ExampleNewHttpClient() {
 
 func ExampleHttpClient_SetCtxHttpClient() {
 	r := &http.Request{} // Here, r is for demo. You should use your handler's request as r.
-	client := Khttp.NewHttpClient(&http.Client{})
+	client := integrations.NewHttpClient(&http.Client{})
 
 	// SetCtxHttpClient is called before mocked http.Client's Get method.
 	client.SetCtxHttpClient(r.Context())
@@ -482,7 +475,7 @@ func ExampleHttpClient_SetCtxHttpClient() {
 
 func ExampleHttpClient_Get() {
 	r := &http.Request{} // Here, r is for demo. You should use your handler's request as r.
-	client := Khttp.NewHttpClient(&http.Client{})
+	client := integrations.NewHttpClient(&http.Client{})
 
 	// SetCtxHttpClient is called before mocked http.Client's Get method.
 	client.SetCtxHttpClient(r.Context())
@@ -497,7 +490,7 @@ func ExampleHttpClient_Get() {
 
 func ExampleHttpClient_Do() {
 	r := &http.Request{} // Here, r is for demo. You should use your handler's request as r.
-	client := Khttp.NewHttpClient(&http.Client{})
+	client := integrations.NewHttpClient(&http.Client{})
 
 	// SetCtxHttpClient is called before mocked http.Client's Do method.
 	client.SetCtxHttpClient(r.Context())
@@ -516,7 +509,7 @@ func ExampleHttpClient_Do() {
 
 func ExampleHttpClient_Post() {
 	r := &http.Request{} // Here, r is for demo. You should use your handler's request as r.
-	client := Khttp.NewHttpClient(&http.Client{})
+	client := integrations.NewHttpClient(&http.Client{})
 
 	// SetCtxHttpClient is called before mocked http.Client's Post method.
 	client.SetCtxHttpClient(r.Context())
