+++ integrations/echo-v4.go
@@ -1,4 +1,4 @@
-package Kecho
+package integrations
 
 import (
 	"bytes"
