+++ integrations/gin-v1.go
@@ -1,4 +1,4 @@
-package Kgin
+package integrations
 
 import (
 	"bufio"
