+++ integrations/mdb_test.go
@@ -1,4 +1,4 @@
-package Kmongo
+package integrations
 
 import (
 	"context"
