+++ integrations/webgo-v6.go
@@ -1,4 +1,4 @@
-package Kwebgo
+package integrations
 
 import (
 	"bytes"
